+++ crates/registry-notary/src/commands/mod.rs
@@ -1,9 +1,11 @@
 mod api_key;
+mod audit;
 mod config_bundle;
 mod healthcheck;
 mod state;
 
 pub(crate) use api_key::*;
+pub(crate) use audit::*;
 pub(crate) use config_bundle::*;
 pub(crate) use healthcheck::*;
 pub(crate) use state::*;